app/requirements.txt
Django
djangorestframework
psycopg2-binary
gunicorn
drf-spectacular
drf-spectacular[sidecar]